decide surface drawing frame rate

+++ rosen/modules/hyper_graphic_manager/frame_rate_manager/hgm_frame_rate_manager.h
@@ -38,7 +38,7 @@ public:
 
 private:
     int GetRefreshRateByScreenId(ScreenId screenId);
-    float GetDrawingFps(int refreshRate, FrameRateRange range);
+    float GetDrawingFps(float refreshRate, FrameRateRange range);
 
     std::unordered_map<ScreenId, FrameRateRange> screenIdToFrameRateRange_;
     std::unordered_map<ScreenId, std::vector<uint32_t>> screenIdToSupportedRefreshRates_;
